Add Heroes with different attack patterns.
We need Heroes to slay the Demon, this commit is a decent start with a
lot of room for improvements.

- The four classes of Hero (Wizard, Ranger, Fighter and Rogue) have
different attack cooldowns.
- Wizards and Rangers have ranged attacks with different projectile
velocities.
- Fighters and Rogues have melee attacks with different strike
distances.
- The melee heroes have a chance of attacking multiple times in quick
succession.
- Rogues also have an alternative ranged attack.

Currently the heroes are static, it would be great to define some
movement patterns for them.

Still need to implement the Hero attacks hitting the Demon and the
Demon taking damage.

Still need to implement the Demon attacks hitting the Heroes and the
Heroes taking damage.